refactor(auth,profile): default Content() for auth + profile blocs
Move AuthenticationScreen, OtpScreen, and ManageProfileScreen into their
public modules; give AuthenticationBloc, OtpBloc, and ManageProfileBloc a
default Content() that delegates to each screen. The impl Blocs and the
preview fakes drop their now-redundant Content overrides.
